refactor: 이모지 삽입 기능을 유틸로 분리하기
